Move AWS EBS operator config to pkg/
It is code and it does not belong to assets.
